New sample with transparent rotating cubes
Added a new sample with the demonstration of the mesh transparency
handling using the Gera WebGL library. There are several cubes which are
mapped with the same texture in scene. Also there is a plane (which is
also rotating as cubes) inside one cube for the depth handling process
demonstration in transparent mode.